Déplacement de l'action slasherattack vers actions.py
